move vite elm plugin into its own folder
