+++ projects/cli/src/effects.js
@@ -1,8 +1,8 @@
 const chokidar = require('chokidar')
 const path = require('path')
 const Vite = require('vite')
-const ElmVitePlugin = require('./vite-plugin/index.js')
-const TypeScriptPlugin = require('./vite-plugin/typescript/index.js')
+const ElmVitePlugin = require('./vite-plugins/elm/index.js')
+const TypeScriptPlugin = require('./vite-plugins/typescript/index.js')
 const { Codegen } = require('./codegen')
 const { Files } = require('./files')
 const { Utils, Terminal } = require('./commands/_utils')
