+++ projects/cli/src/vite-plugins/elm/README.md
@@ -1,0 +1,3 @@
+# @elm-land/vite-plugin-elm
+> A plugin that shows full-color Elm compiler errors in the HMR overlay
+
